requirements.txt
# Core dependencies
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0
astropy>=5.0.0
pyyaml>=6.0

# Machine learning
tensorflow>=2.10.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
tqdm>=4.62.0

# File I/O
h5py>=3.6.0

# Parallel processing
joblib>=1.1.0

